refactor: App.js를 pages 레이어로 분리
